+++ trunk/src/interception/nl_firewall.c
@@ -11,7 +11,7 @@ int nl_firewall_init(){
 
 static char buffer[65536];
 
-struct iphdr *nl_firewall_recv(int sock,int *packet_id){
+struct iphdr *nl_firewall_recv(int sock,unsigned long *packet_id){
 	nl_recv(sock,buffer,sizeof(buffer));
 	struct ipq_packet_msg *msg = nl_payload(buffer);
 	*packet_id=msg->packet_id;
